Split endpoint from IPC
