+++ kernel/domain/syscalls/create_endpoint.c
@@ -31,7 +31,7 @@
 
 #include <jinue/shared/asm/errno.h>
 #include <kernel/descriptor.h>
-#include <kernel/ipc.h>
+#include <kernel/endpoint.h>
 #include <kernel/object.h>
 #include <kernel/process.h>
